+++ app-frame/macrostrat/app_frame/__init__.py
@@ -1,4 +1,6 @@
-from .compose import compose
+from macrostrat.app_frame.compose.base import compose
+
+from .control_command import BackendType, CommandBase, ControlCommand
 from .core import Application
 from .exc import ApplicationError
 from .subsystems import Subsystem, SubsystemError, SubsystemManager