+++ utils/macrostrat/utils/__init__.py
@@ -6,6 +6,7 @@ import os
 from contextlib import contextmanager
 from pathlib import Path
 
+from .exc import ApplicationError, BaseError
 from .logs import get_logger, setup_stderr_logs
 from .shell import cmd, split_args
 
@@ -25,3 +26,15 @@ def working_directory(path: Path):
     os.chdir(str(path))
     yield
     os.chdir(prev_cwd)
+
+
+@contextmanager
+def override_environment(**kwargs):
+    """Override environment variables for a block of code."""
+    old_environ = dict(os.environ)
+    os.environ.update(kwargs)
+    try:
+        yield
+    finally:
+        os.environ.clear()
+        os.environ.update(old_environ)